Dev: Start work on restructuring mod preset system
